django_project_root/requirements.txt
alembic==1.4.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.4
django-crispy-forms==1.10.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
dominate==2.6.0
fake2db==0.5.4
Faker==0.7.11
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
numpy==1.19.4
phonenumbers==8.12.16
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
virtualenv==20.2.1
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
